// Frontend/FrontendSegurosABC/src/app/src/app/api/model/segurosAbc.ts
/**
 * BackendSegurosABC
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface SegurosAbc { 
    id?: number;
    identificacion?: number;
    primerNombre?: string | null;
    segundoNombre?: string | null;
    primerApellido?: string | null;
    segundoApellido?: string | null;
    telefono?: string | null;
    correoElectronico?: string | null;
    fechaNacimiento?: string;
    valorEstimado?: number;
    observaciones?: string | null;
}